Per Liedman's updated 3DS loader.  Here's his notes on this new version:
I have done some work on the 3DS loader for SSG the last days. Here's some info
on what I've done:

* The loader no longer loads the whole file to memory, but rather loads data
from the file as it needs it. This should (hopefully) fix problems that some
64-bit systems had with the loader (I seem to remember that someone had
problems since the loader accessed data that wasn't 4-byte aligned - this
shouldn't be a problem any longer, although I have no way to test it).

* It now uses the ssgCreateData structure to create leaf-nodes. As I
understand it, there's no way of telling the ssgCreateData structure that a
texture shouldn't be warped (that it should use GL_CLAMP), so this
functionality has been removed from the loader, although it should be a
breeze to plug it in again if ssgCreateData supports this.

* Errors are now raised by using ulSetError instead of printf.

* All material parameters are now set to default values (I think?!), so the
problem with GL_EMISSION getting the wrong values should be history.

* Added a mechanism to debug-output the format of the file (define DEBUG to
use this). Might not only be usefull for debugging, but also to get a
better picture of how 3DS files are structured.

* Cleaned up a lot of old stuff.

I've tested the loader with around 10-20 models, and regretfully I must say
that there are still some issues with it:

* Some models (one of the test cases) gets turned "inside out" - obviously the
triangle winding has been reversed so that all triangles that should be
culled is shown and vice versa. I don't know where this information can
be found in the file.

* Models that uses double-sided materials sometimes look strange. My approach
to double sided materials is to add a copy of each triangle with the winding
reversed and also negate the normals...this seems to *almost* work...any
ideas?

* The transformation matrix in the 3DS files is still a mystery to me -
whatever I do with it seems to make things worse than ignoring it, although
that doesn't work perfect either. I've taken a look at other loaders, and it
seems that no one really knows what to do with them.


git-svn-id: https://svn.code.sf.net/p/plib/code/trunk@156 713b2cfd-b11f-0410-b270-d5401603882f