There are two removeKid() functions, one by index and one by *kid_ptr.
The removeKid(index) function is just a wrapper that looks up the kid
pointer and calls removeKid(kid_ptr *)

However removeKid(by kid pointer *) always does a linear search to
find the kid index again.

This patch simply reverses which is the wrapper and which is the actual
deleting function.

I assert that it makes more sense to have removeKid(index) do the
actual delete because it already knows the index and can do the delete
directly.

removeKid(kid_ptr *) can be the wrapper which searches for an appropriate
index and calls removeKid(index)


git-svn-id: https://svn.code.sf.net/p/plib/code/trunk@1720 713b2cfd-b11f-0410-b270-d5401603882f